chrono: reorganise directories following upstream
