+++ HLTrigger/Timer/test/chrono/src/mach_absolute_time.cc
@@ -8,7 +8,7 @@
 #include <mach/mach.h>
 #include <mach/mach_time.h>
 
-#include "mach_absolute_time.h"
+#include "interface/mach_absolute_time.h"
 
 // read the calibration of mach_absolute_time
 static
