+++ HLTrigger/Timer/test/chrono/src/posix_clock_gettime.cc
@@ -1,4 +1,4 @@
-#include "posix_clock_gettime.h"
+#include "interface/posix_clock_gettime.h"
 
 #include <unistd.h>
