+++ HLTrigger/Timer/test/chrono/src/tbb_tick_count.cc
@@ -1,3 +1,3 @@
-#include "tbb_tick_count.h"
+#include "interface/tbb_tick_count.h"
 
 const tbb::tick_count clock_tbb_tick_count::epoch = tbb::tick_count::now();
